requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
ollama==0.4.4
sqlalchemy==2.0.35
pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.2
jinja2==3.1.4
python-dotenv==1.0.1
aiofiles==24.1.0
